fix(sunabar): widen TransactionList.Count to int64 to fully eliminate narrowing (CWE-190)
`int(count)` の narrowing を残したままだと CodeQL の go/incorrect-integer-conversion
は data flow が int への conversion に到達するため open のまま残った。
TransactionList.Count を int64 に格上げして narrowing そのものを排除する。
件数の負値だけは引き続き API 異常として弾く。 大きな値を受理することを
TestListTransactions_AcceptsLargeCount で固定。

+++ internal/platform/sunabar/types.go
@@ -41,13 +41,15 @@ type ListTransactionsParams struct {
 }
 
 // TransactionList 入出金明細の一覧。
+// Count は sunabar 実 API が返す件数 ( 64bit) をそのまま保持する。
+// int への narrowing は避け、 32bit 環境でのオーバーフローを防ぐ。
 type TransactionList struct {
 	AccountID    string
 	CurrencyCode string
 	BaseDate     string
 	BaseTime     string
 	HasNext      bool
-	Count        int
+	Count        int64
 	Transactions []Transaction
 }
